Try to reduce APK for API29+
